requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-filter==23.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.3
phonenumbers==8.13.8
phonenumberslite==8.13.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
zipp==3.15.0
